+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Remove three dead rendezvous points
+
 ## [4.7.10] - 2026-06-02
 
 ## [4.7.9] - 2026-05-29