+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.8.3] - 2026-06-08
 
+- Reliably retry Monero redeem step
+
 ## [4.8.2] - 2026-06-06
 
 ## [4.8.1] - 2026-06-05